+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.0.2 -- *work in progress*
 * Use `data-` naming in attributes passed
 * Pass the DOM element as `this` to the init function
+* An init function and be defined on any element
 
 
 ## v0.0.1 -- 24 Jan 2015